requirements.txt
django
whitenoise
djangorestframework

